+++ modules/client/src/internal/interfaces.ts
@@ -78,6 +78,10 @@ export interface IClientMethods {
   prepareInstallation: (
     params: PrepareInstallationParams,
   ) => AsyncGenerator<PrepareInstallationStepValue>;
+    // gets the protol version of a dao
+  getProtocolVersion: (
+    contractAddress: string,
+  ) => Promise<[number, number, number]>;
 }
 
 export interface IClientEncoding {