+++ modules/client/CHANGELOG.md
@@ -20,6 +20,7 @@ TEMPLATE:
 ### Added
 - Added `initializeFrom` encoders and decoders 
 - Support for ERC721 deposits and withdrawals
+- Added `getProtocolVersion` function
 ### Fixes
 - Fix status calculation for token voting proposals
 - Make the `network` parameter required on `getPluginInstallItem`